feat(types): add auth, login, and http fetch types
